refactor(server): migrate transaction amount fields from float to str/Decimal
Float loses precision for monetary values. The codebase already used
str + Decimal validators for recurring transactions and search filters;
this finishes the migration on the record/create path:

- CreateTransactionItem.amount, TransactionItem.amount (LangGraph tool),
  ExecuteTransferInput.amount: float -> str with a @field_validator that
  parses Decimal, rejects non-positive, normalizes to 8 decimals.
- TransactionDisplayValue.from_params: accept float | str | Decimal
  (backward compatible with existing float callers), format via Decimal.
- create_transaction / update_transaction service signatures
  (crud + facade + create_transaction_for_space): float -> Decimal.
  Drop the redundant Decimal(str(amount)) conversions; input is now
  typed Decimal.
- update_transaction: amount <= 0 now raises BusinessError instead of
  being silently ignored.
- record_transactions summary totals: sum via Decimal, serialize to
  float for the GenUI payload.
- execute_transfer: convert str -> Decimal at the service boundary.

Test data updated to pass amounts as strings/Decimal. 120 unit+integration
tests pass.

+++ server/app/services/shared_space_service.py
@@ -614,7 +614,7 @@ class SharedSpaceService:
         self,
         user_uuid: UUID,
         space_id: UUID,
-        amount: float,
+        amount: Decimal,
         transaction_type: str = "expense",
         transaction_at: datetime | None = None,
         category_key: str = "OTHERS",
